FGVC_PIM/requirements.txt
matplotlib==3.3.1
numpy
opencv-python==4.5.2.54
pillow==8.2.0
pip==21.1.3
seaborn==0.11.0
torch
torchvision
wandb==0.12.4
